Translate import formats via /web
If the remote URL is text/plain or one of the known content types for
the formats we support (e.g., application/x-bibtex), try to handle it
via import translation. This allows, for example, a remote BibTeX file
to be translated.

This adapts some of the code from #59 in a more general way, with a
responseTypeMap property that can be passed to HTTP.request().