Expose attachment urls via API
Generally, it is not useful to expose regular Zotero attachments via
the API. However, url attachments are an exception: url resources are
of use to the api user, as they help locating various linked online
artifacts that can be readily accessed (in contrast to other attachment
types residing within Zotero's store). Therefore, this commit exports
them as linked items into the API-response.