Re-apply "Update Zotero to node 18"
This reverts commit 84345cd5286dcaba82cb9eb8c39c928688fc92a4.

Reason for revert: Try this again post mesh upgrades

Previously upgrading to 18 failed, however it is possible this
was due to an outdated mesh configuration which wasn't compatible
with node 18. We should try again and see if this is now fixed.

Mesh was updated in Iba98da4859cc645a4d

Bug: T361728
Change-Id: I242b332ed65466ad67137ce3c90b5945ef65b558