Do not throw in HTTP.request().then() handler for errors
HTTP.doGet() and HTTP.doPost() are not async functions and the promise
from HTTP.request() is not awaited for, so any throws in the then()
handler are not handled anywhere. This is likely causing failures to
fallback to a different translator when a non-success response code is
received for a request from a translator in certain deployment
conditions.

Addresses #178